refactor(pi): inline resolveModel into session-factory
